refactor: consolidate small tool files (14 → 10)
Merge sandbox.py into sandbox_common.py, volume_helpers.py into
modal_volumes.py, runtime_module_helpers.py into llm_tools.py.
Clean up __init__.py re-exports.
