+++ src/fleet_rlm/runtime/tools/sandbox_delegate_tools.py
@@ -10,11 +10,11 @@ from fleet_rlm.runtime.agent.recursive_runtime import spawn_delegate_sub_agent_a
 from fleet_rlm.runtime.agent.signatures import GroundedCitation
 from fleet_rlm.runtime.agent.tool_delegation import _sync_compatible_tool_callable
 
-from .runtime_module_helpers import coerce_int as _coerce_int
-from .runtime_module_helpers import coerce_str_list as _coerce_str_list
-from .runtime_module_helpers import prediction_value as _prediction_value
-from .runtime_module_helpers import run_cached_runtime_module as _run_runtime_module
-from .runtime_module_helpers import runtime_metadata as _runtime_metadata
+from .llm_tools import coerce_int as _coerce_int
+from .llm_tools import coerce_str_list as _coerce_str_list
+from .llm_tools import prediction_value as _prediction_value
+from .llm_tools import run_cached_runtime_module as _run_runtime_module
+from .llm_tools import runtime_metadata as _runtime_metadata
 from .sandbox_common import _aexecute_submit_ctx, _SandboxToolContext
 from .shared import (
     build_trajectory_payload,
